feat(cli-split): split the 12 hybrid commands into *_local.go + local_stubs_client.go (#1775)
For create, delete, get, list, info, label list/set/remove, install-stack,
resize: move every local-mode-only function into a sibling *_local.go
tagged !containarium_client (8 files, matching the design doc exactly:
create_local, delete_local, get_local, list_local, info_local,
label_local, install_stack_local, resize_local). ssh-config sync and
prune need no file of their own — they already dispatch through
listLocal()/deleteLocal().

Beyond the functions that already existed as standalone (createLocal,
deleteLocal, getLocal, listLocal, getLabelsLocal/setLabelsLocal/
removeLabelsLocal, installStackLocal, runResizeLocal — moved verbatim),
four things had to be extracted from inline serverAddr == "" branches to
fully satisfy "create.go/delete.go import neither pkg/core/container nor
call incus.New" under the client build:

- containerExistsLocal (create's --force existence check)
- createJumpServerAccountLocal / cleanupJumpServerAccountLocal (create's
  host-account create + on-failure cleanup)
- deleteJumpServerAccountLocal (delete's host-account cleanup)
- getSystemInfoLocal / getContainerInfoLocal ("the two info helpers" the
  design's stub list names generically)

Each became its own small function so behavior stays byte-for-byte
identical under containariumd — same call sites, same order, same error
text — rather than literally merging into createLocal/deleteLocal's
bodies (which would have reordered stdout: the account-creation/cleanup
prints happen at specific points relative to the container-create prints
today, and "byte-for-byte unchanged" is this issue's own done-when).

local_stubs_client.go (containarium_client) defines all 15 of the above
as one-liners returning a shared errNoLocalMode ("this is the
containarium client; it has no local mode — pass --server, run
`containarium login`, or use containariumd on the host") — 9 named in
the design doc's stub list plus the 6 extras above.

Done-when, verified:
- containariumd behavior in local mode is byte-for-byte unchanged: full
  repo test suite (go test -race, minus test/integration) green under
  the default tag set
- Under containarium_client, create.go/delete.go import neither
  pkg/core/container nor incus.New (grep-verified)
- Each stub's error names --server, containarium login, and
  containariumd
- New table-driven test (local_stubs_client_test.go): all twelve hybrid
  commands, invoked with no server anywhere, return an error satisfying
  errors.Is(err, errNoLocalMode) — i.e. every one resolves to a stub,
  never to working Incus code. Redirects $HOME to an empty temp dir so
  the assertion doesn't depend on (or get broken by) whichever operator
  is actually logged in on the machine running the test — `info`'s
  isCloudTarget check reads ~/.containarium/credentials.json, and this
  dev box has a real one.
- go build -tags containarium_client ./cmd/containarium compiles; go
  build/vet/test clean under both tag sets; golangci-lint 0 issues

Design: docs/architecture/cli-client-server-split.md (PR #1769). Part
of #1787 (Sprint 2026-W38). Builds on #1773 and #1774.
