+++ internal/cmd/get_local.go
@@ -1,0 +1,18 @@
+//go:build !containarium_client
+
+package cmd
+
+import (
+	"fmt"
+
+	"github.com/footprintai/containarium/pkg/core/container"
+	"github.com/footprintai/containarium/pkg/core/incus"
+)
+
+func getLocal(username string) (*incus.ContainerInfo, error) {
+	mgr, err := container.New()
+	if err != nil {
+		return nil, fmt.Errorf("failed to connect to Incus: %w (is Incus running?)", err)
+	}
+	return mgr.Get(username)
+}
