+++ internal/cmd/label_remove.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 
 	"github.com/footprintai/containarium/internal/client"
-	"github.com/footprintai/containarium/pkg/core/container"
 	"github.com/spf13/cobra"
 )
 
@@ -42,49 +41,6 @@ func runLabelRemove(cmd *cobra.Command, args []string) error {
 	}
 
 	return removeLabelsLocal(username, keys)
-}
-
-func removeLabelsLocal(username string, keys []string) error {
-	mgr, err := container.New()
-	if err != nil {
-		return fmt.Errorf("failed to connect to Incus: %w (is Incus running?)", err)
-	}
-
-	containerName := username + "-container"
-
-	// Check if container exists
-	if !mgr.ContainerExists(containerName) {
-		return fmt.Errorf("container %q does not exist", containerName)
-	}
-
-	// Get current labels to show which were actually removed
-	// Note: Manager methods expect username (without -container suffix)
-	currentLabels, err := mgr.GetLabels(username)
-	if err != nil {
-		return fmt.Errorf("failed to get current labels: %w", err)
-	}
-
-	removedCount := 0
-	for _, key := range keys {
-		if _, exists := currentLabels[key]; exists {
-			if err := mgr.RemoveLabel(username, key); err != nil {
-				return fmt.Errorf("failed to remove label %q: %w", key, err)
-			}
-			fmt.Printf("Removed label: %s\n", key)
-			removedCount++
-		} else {
-			if verbose {
-				fmt.Printf("Label not found: %s (skipped)\n", key)
-			}
-		}
-	}
-
-	if removedCount > 0 {
-		fmt.Printf("\nRemoved %d label(s) from container %s\n", removedCount, containerName)
-	} else {
-		fmt.Println("No labels were removed (none matched)")
-	}
-	return nil
 }
 
 func removeLabelsRemoteHTTP(username string, keys []string) error {
