+++ internal/cmd/label_set.go
@@ -5,7 +5,6 @@ import (
 	"strings"
 
 	"github.com/footprintai/containarium/internal/client"
-	"github.com/footprintai/containarium/pkg/core/container"
 	"github.com/spf13/cobra"
 )
 
@@ -68,52 +67,6 @@ func runLabelSet(cmd *cobra.Command, args []string) error {
 	}
 
 	return setLabelsLocal(username, labels)
-}
-
-func setLabelsLocal(username string, labels map[string]string) error {
-	mgr, err := container.New()
-	if err != nil {
-		return fmt.Errorf("failed to connect to Incus: %w (is Incus running?)", err)
-	}
-
-	containerName := username + "-container"
-
-	// Check if container exists
-	if !mgr.ContainerExists(containerName) {
-		return fmt.Errorf("container %q does not exist", containerName)
-	}
-
-	// Get current labels if not overwriting
-	// Note: Manager methods expect username (without -container suffix)
-	if !labelOverwrite {
-		currentLabels, err := mgr.GetLabels(username)
-		if err != nil {
-			return fmt.Errorf("failed to get current labels: %w", err)
-		}
-		// Only set labels that don't already exist
-		for key, value := range labels {
-			if _, exists := currentLabels[key]; !exists {
-				if err := mgr.AddLabel(username, key, value); err != nil {
-					return fmt.Errorf("failed to add label %s=%s: %w", key, value, err)
-				}
-				fmt.Printf("Added label: %s=%s\n", key, value)
-			} else {
-				fmt.Printf("Skipped existing label: %s (use --overwrite to update)\n", key)
-			}
-		}
-		return nil
-	}
-
-	// Set all labels (overwriting existing ones)
-	for key, value := range labels {
-		if err := mgr.AddLabel(username, key, value); err != nil {
-			return fmt.Errorf("failed to set label %s=%s: %w", key, value, err)
-		}
-		fmt.Printf("Set label: %s=%s\n", key, value)
-	}
-
-	fmt.Printf("\nLabels set on container %s\n", containerName)
-	return nil
 }
 
 func setLabelsRemoteHTTP(username string, labels map[string]string) error {
